chore(nvim): test role with molecule
Quite bumpy road. Molecule documentation is BAD about how roles
should be tested inside collection.
So i am trying to fix the role and the molecule testing at the
same time. Somewhat got it partially working, but there is a lot
impromevents to make. Packer is not configured automatically.
TreeSitter parsers is not instlled and so on.
One thing i liked is that i figured out how to beautify ansible
outpur. Look into molecule.yml for specific settings.